+++ microservice-core/include/settings/IServerSettings.hpp
@@ -2,16 +2,14 @@
 
 #include <string>
 #include <cstddef>
+#include <chrono>
 
-/**
- * @file IServerSettings.hpp
- * @brief Интерфейс настроек сервера
- * @author Anton Tobolkin
- */
 class IServerSettings {
 public:
     virtual ~IServerSettings() = default;
     virtual std::string getHost() const = 0;
     virtual int getPort() const = 0;
     virtual size_t getMaxRequestBodySize() const = 0;
+    virtual std::chrono::milliseconds getReadTimeout() const = 0;
+    virtual std::chrono::milliseconds getWriteTimeout() const = 0;
 };